Cherry pick #3953 #3955 (#3960)
* Cleanup URs on trigger deletion (#3955)

* Clean URs on trigger deletion


* Make kyverno api import aliases consistent


* Fix gofumpt error


* Remove unused code



* fix panic issue for ur (#3953)

* fix the import
